+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Added
+- Added `as_arrival_curve_prefix()` to the arrival-model interface to obtain arrival-curve prefixes, plus `Task` and `TaskSet` helpers to convert arrivals to arrival-curve prefixes.
+
 ## [0.1.0] - 2026-01-13
 
 First public release